Move pack creation into a shared place.
